Beginings of the Virtual Memory Manager.
This commit begins introducing a basic virtual memory manager
into the Kernel. It needs to adopt a number of structures and
page tables left behind by the Boot Loader. This slightly
complicates the implementation as it needs to work in an
environment that does not have immediate access to all physical
memory. Further more the bootloader _may_ have setup the kernel
as a higher half kernel (depending on the kernel ELF) and thus
identity mapping structures is not an option.