Add the concept of a frontmost process.
This will be required for eventually working out where to direct
user input in the system. Currently this is implemented with a
couple of shortcuts.

Further, this comments out some of the logging information for the
kernel heap as it is just too muvh and makes it difficult to debug
other things.