WIP (Broken): Begin adding a device manager.
This is the first part of what will likely be a large number of
commits, making a large number of changes. This removes the
old serial port driver and replaces it with an RS232 driver that
can be made aware of the various COM ports of the CPU. It is
currently primitive but it should help provide a more solid
device/driver interface going forwards.