Strip out old "terminal" layer and begin adding VT100 emulation layer.
This is the beginning of some major work on the monitor/terminal/screen
and a continuation of the device manager work. The VT100 terminal
device is now responsible for driving the terminal output. It currently
is incapable of driving a VESA display without doing some strange things
but it lays the initial ground work and works correctly with a VGA text
mode.