Introduce 'keyboard' process to handle keyboard input.
The PS/2 keyboard interrupt now sends data over a pipe to the
keyboard process which handles parsing it and then sending the data
via a pipe to the current key process (currently always the kernel).

This also introduces the getc(FILE *) function which can be used to
wait for key events in a given process.