Use leave instruction. Fix builds.
Inspection of GCC output reveals the use of the 'leave' instruction
in place of the following:

	mov esp, ebp
	pop ebp

This should probably be favoured in CoreLoader as well to reduce
amount of explicit instructions and receive internal CPU
optimisations found in the instruction.

However the counterpart 'enter' instruction is _not_ used by GCC.
Research on this reveals that the 'enter' instruction has some
major performance issues.